[Green] Task №7

+++ LaboratoryWork1/histogram.xaml.cs
@@ -26,7 +26,7 @@ namespace LaboratoryWork1
 
             for (int index = 0; index < myAL.Count; index++)
             {
-                chart.Series["Series1"].Points.AddXY(index + 1, (int)myAL[index]);
+                chart.Series["Series1"].Points.AddXY(index + 1, myAL[index]);
             }
         }
     }
